pj: fix `/place` research endpoint
